requirements.txt
pygame==2.1.2
torch==2.0.0
numpy==1.23.0
python-chess==1.999
matplotlib==3.5.1
